Web Server Implementation
Adds a basic web interface that reflects the state of the door and lock in real time.

The basic design is to have a small static index.html with all assets inlined so we only need a
single HTTP endpoint, this index.html is embedded in the binary at buildtime.

When the client downloads the index.html, included javascript will open a websocket through which
state updates will be received, and lock/unlock commands can be sent.